리뷰 후 수정(PokerGame 수정)
 - start, shareCards, registerPlayer를 PokerGame에서 구현
 - Dealer 추후 점수 계산을 위해 calculateScore() 구현
